+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changes
 0.2 (unreleased)
 ----------------
 
+- Replace deprecated use of ``Adopt`` by ``MappingAdopt``.
+  [rnix]
+
 - ``node.ext.yaml.YamlNode`` and ``node.ext.yaml.YamlFile`` not provides a
   default child factory any more.
   [rnix]